move knowledge base service to unique toolkit folder
