+++ unique_toolkit/unique_toolkit/__init__.py
@@ -2,7 +2,7 @@
 from unique_toolkit.chat import ChatService
 from unique_toolkit.content import ContentService
 from unique_toolkit.embedding import EmbeddingService
-from unique_toolkit.knowledge_base.service import KnowledgeBaseService
+from unique_toolkit.knowledge_base import KnowledgeBaseService
 from unique_toolkit.language_model import (
     LanguageModelMessages,
     LanguageModelName,
